trading_agents_adk/pyproject.toml
[project]
name = "trading-agents-adk"
version = "0.1.0"
description = "Multi-Agent LLM Trading Framework built on Google ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
trading-agents = "main:_sync_main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools", "graph", "dataflows"]
